ARM: mpu: add MPU initialisation for secondary cores
The MPU initialisation on the primary core is performed in two stages,
one
minimal stage to ensure the CPU can boot and a second one after
sanity_check_meminfo. As the memory configuration is known by the time
we
boot secondary cores only a single step is necessary, provided the
values
for DRSR are passed to secondaries.

This patch implements this arrangement. The configuration generated for
the
MPU regions is made available to the secondary core, which can then use
the
asm MPU intialisation code to program a complete region configuration.

This is necessary for SMP configurations without an MMU, as the MPU
initialisation is the only way to ensure that memory is specified as
'shared'.

Signed-off-by: Jonathan Austin <jonathan.austin@arm.com>
Reviewed-by: Will Deacon <will.deacon@arm.com>
CC: Nicolas Pitre <nico@linaro.org>
Signed-off-by: Jake Weinstein <xboxlover360@gmail.com>

Conflicts:
	arch/arm/kernel/smp.c