cpufreq: interactive: don't skip waking up speedchange_task if target…
…_freq > policy->cur

When __cpufreq_driver_target() in speedchange_task failed for some
reason, the
policy->cur could be lower than the target_freq. The governor misses to
change
the target_freq if the target_freq is equal to the next_freq at the next
sample
time.

Added a check to prevent the CPU to stay at the speed that is lower than
the
target_freq for long duration.

Change-Id: Ibfdcd193b8280390b8f8374a63218aa31267f310
Signed-off-by: Minsung Kim <ms925.kim@samsung.com>